Overhaul KDE config, allow for no home partition
Previously, the KDE install would result in the plasma-desktop meta-package plus a bunch of other individual packages. Changed to just use the broader kde5 and kde5-baseapps meta-packages. This also means that SDDM is used as a DM rather than emptty, which in turn means that Xorg is required even for a KDE Plasma Wayland install.

Also added an option for not having a separate home partition.